# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Quotation App
  version: 2.0.0
  description: Multi-tenant SaaS quotation & CRM platform for MSMEs
servers:
  - url: /api
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      operationId: listTasks
      parameters:
        - in: query
          name: status
          schema: { type: string }
        - in: query
          name: dueWithinDays
          schema: { type: integer }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Task" }
    post:
      tags: [Tasks]
      operationId: createTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
  /tasks/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    patch:
      tags: [Tasks]
      operationId: updateTask
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TaskInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Task" }
    delete:
      tags: [Tasks]
      operationId: deleteTask
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MessageResponse" }

  # ── Calls ──
...
  /sales-orders:
    get:
      tags: [SalesOrders]
      operationId: listSalesOrders
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SalesOrder" }
    post:
      tags: [SalesOrders]
      operationId: createSalesOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalesOrderInput" }
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
  /sales-orders/{id}:
    parameters:
      - in: path
        name: id
        required: true
        schema: { type: integer }
    get:
      tags: [SalesOrders]
      operationId: getSalesOrder
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrderDetail" }
    patch:
      tags: [SalesOrders]
      operationId: updateSalesOrder
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SalesOrderInput" }
      responses:
        "200":
          description: ok
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }
  /sales-orders/from-quotation/{quotationId}:
    parameters:
      - in: path
        name: quotationId
        required: true
        schema: { type: integer }
    post:
      tags: [SalesOrders]
      summary: Promote quotation to sales order
      operationId: promoteQuotationToSalesOrder
      responses:
        "201":
          description: created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SalesOrder" }

  # ── Invoices ──
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message: { type: string }

    # ── Auth ──
...
    Task:
      type: object
      required: [id, title, status, priority, createdAt]
      properties:
        id: { type: integer }
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string }
        priority: { type: string }
        dueAt: { type: ["string", "null"] }
        relatedType: { type: string }
        relatedId: { type: ["integer", "null"] }
        assignedToId: { type: ["integer", "null"] }
        assignedToName: { type: ["string", "null"] }
        completedAt: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    TaskInput:
      type: object
      required: [title]
      properties:
        title: { type: string }
        description: { type: ["string", "null"] }
        status: { type: string, enum: [open, done, cancelled] }
        priority: { type: string, enum: [low, medium, high] }
        dueAt: { type: ["string", "null"] }
        relatedType: { type: string, enum: [lead, client, quotation, invoice, none] }
        relatedId: { type: ["integer", "null"] }
        assignedToId: { type: ["integer", "null"] }
...
    SalesOrder:
      type: object
      required: [id, orderNumber, status, subtotal, total, createdAt]
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        clientId: { type: ["integer", "null"] }
        clientName: { type: ["string", "null"] }
        quotationId: { type: ["integer", "null"] }
        status: { type: string }
        subtotal: { type: number }
        discountAmount: { type: number }
        taxAmount: { type: number }
        total: { type: number }
        expectedDeliveryAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string }
        updatedAt: { type: string }
    SalesOrderItem:
      type: object
      required: [id, description, quantity, unitPrice, totalPrice]
      properties:
        id: { type: integer }
        description: { type: string }
        quantity: { type: integer }
        unitPrice: { type: number }
        totalPrice: { type: number }
    SalesOrderDetail:
      allOf:
        - $ref: "#/components/schemas/SalesOrder"
        - type: object
          properties:
            items:
              type: array
              items: { $ref: "#/components/schemas/SalesOrderItem" }
    SalesOrderInput:
      type: object
      properties:
        clientId: { type: ["integer", "null"] }
        status: { type: string, enum: [draft, confirmed, in_production, delivered, cancelled] }
        expectedDeliveryAt: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        items:
          type: array
          items:
            type: object
            required: [description, quantity, unitPrice]
            properties:
              description: { type: string }
              quantity: { type: integer }
              unitPrice: { type: number }
...