# lib/api-zod/src/generated/types/salesOrder.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Multi-tenant SaaS quotation & CRM platform for MSMEs
 * OpenAPI spec version: 2.0.0
 */

export interface SalesOrder {
  id: number;
  orderNumber: string;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  quotationId?: number | null;
  status: string;
  subtotal: number;
  discountAmount?: number;
  taxAmount?: number;
  total: number;
  /** @nullable */
  expectedDeliveryAt?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}